requirements.txt
asn1crypto==1.5.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
filelock==3.15.4
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.1
requests==2.32.3
snowflake-connector-python==3.12.0
sortedcontainers==2.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
